requirements.txt
bootstrap-flask
Flask-WTF
openai

